backend/requirements.txt
fastapi==0.111.0         # latest stable
uvicorn[standard]==0.29.0
itsdangerous>=2.1.2
pydantic>=2.0
pydantic-settings>=2.0
pandas
statsmodels
prophet
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
python-multipart>=0.0.5
boto3>=1.26.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-jose[cryptography]>=3.3.0
python-dotenv==1.0.1
bcrypt==4.1.2            # pin so passlib can read __about__.__version__
